feat: move analyze to plugins
