+++ packages/solutions/app-tools/src/plugins/analyze/constants.ts
