+++ packages/solutions/app-tools/src/plugins/analyze/getClientRoutes/getRoutes.ts
