+++ packages/solutions/app-tools/src/plugins/analyze/getClientRoutes/getRoutesLegacy.ts
