+++ packages/solutions/app-tools/src/plugins/analyze/getClientRoutes/index.ts
