+++ packages/solutions/app-tools/src/plugins/analyze/getFileSystemEntry.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import { findExists, ensureAbsolutePath } from '@modern-js/utils';
 import type { Entrypoint } from '@modern-js/types';
-import type { AppNormalizedConfig, IAppContext } from '../types';
+import type { AppNormalizedConfig, IAppContext } from '../../types';
 import { isDefaultExportFunction } from './isDefaultExportFunction';
 import {
   JS_EXTENSIONS,
