+++ packages/solutions/app-tools/src/plugins/analyze/makeLegalIdentifier.ts
